Move DeltaMptIterator into mod delta_mpt.
