+++ core/src/storage/impls/storage_manager/storage_manager.rs
@@ -842,45 +842,6 @@ lazy_static! {
     );
 }
 
-// FIXME: move to another file.
-#[derive(Clone)]
-pub struct DeltaMptIterator {
-    pub mpt: Arc<DeltaMpt>,
-    pub maybe_root_node: Option<NodeRefDeltaMpt>,
-}
-
-impl DeltaMptIterator {
-    pub fn iterate<'a, DeltaMptDumper: KVInserter<(Vec<u8>, Box<[u8]>)>>(
-        &self, dumper: &mut DeltaMptDumper,
-    ) -> Result<()> {
-        match &self.maybe_root_node {
-            None => {}
-            Some(root_node) => {
-                let db = &mut *self.mpt.db_owned_read()?;
-                let owned_node_set = Default::default();
-                let mut cow_root_node =
-                    CowNodeRef::new(root_node.clone(), &owned_node_set);
-                let guarded_trie_node =
-                    GuardedValue::take(cow_root_node.get_trie_node(
-                        self.mpt.get_node_memory_manager(),
-                        &self.mpt.get_node_memory_manager().get_allocator(),
-                        db,
-                    )?);
-                cow_root_node.iterate_internal(
-                    &owned_node_set,
-                    &self.mpt,
-                    guarded_trie_node,
-                    CompressedPathRaw::new_zeroed(0, 0),
-                    dumper,
-                    db,
-                )?;
-            }
-        }
-
-        Ok(())
-    }
-}
-
 use super::{
     super::{
         super::{
@@ -890,12 +851,11 @@ use super::{
                 delta_db_manager::*, snapshot_db::*,
                 snapshot_db_manager::SnapshotDbManagerTrait,
             },
-            utils::{arc_ext::*, guarded_value::GuardedValue},
+            utils::arc_ext::*,
             StateRootWithAuxInfo,
         },
         delta_mpt::*,
         errors::*,
-        merkle_patricia_trie::{CompressedPathRaw, KVInserter},
         state_manager::*,
     },
     *,
