+++ core/src/storage/mod.rs
@@ -80,7 +80,6 @@ pub use self::{
             snapshot_db_manager_sqlite::SnapshotDbManagerSqlite,
             sqlite::SqliteConnection,
         },
-        storage_manager::DeltaMptIterator,
     },
     state::{State as StorageState, StateTrait as StorageStateTrait},
     state_manager::{
